UniInject: moved tests to own namespace
